requirements.txt
neptune-client==0.16.16
nflows==0.14
numba==0.56.4
scikit-dimension==0.3
scikit-learn==1.2.1
sklearn==0.0
torch==1.13.1
tqdm==4.64.1
numpy==1.21.6
torchdiffeq=0.2.3